Fix validation of partial files
Because the exact date of the header in the single ZIP archive is
different now, some tests are broken. This needs to be fixed in a future commit.
